Decompose GlobalTab into ScreenshotPanel and RssAlertPanel (closes #71)
