+++ src/renderer/src/contacts/ScreenshotPanel.css
@@ -1,0 +1,103 @@
+.sp-grid {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  margin-top: 6px;
+}
+
+.sp-thumb {
+  position: relative;
+  width: 80px;
+  height: 56px;
+  overflow: hidden;
+  border: 1px solid var(--color-border);
+  cursor: pointer;
+  background: var(--color-bg);
+}
+
+.sp-thumb-error {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 11px;
+  color: var(--color-danger);
+  padding: 0 4px;
+  text-align: center;
+}
+
+.sp-thumb-img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.sp-thumb-loading {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 18px;
+  color: var(--color-text-muted);
+}
+
+.sp-delete-confirm {
+  position: absolute;
+  inset: 0;
+  background: rgba(0, 0, 0, 0.75);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 4px;
+}
+
+.sp-delete-label {
+  font-size: 10px;
+  color: #fff;
+  font-weight: 600;
+  font-family: var(--font-mono);
+}
+
+.sp-delete-actions {
+  display: flex;
+  gap: 4px;
+}
+
+.sp-delete-yes {
+  font-size: 10px;
+  padding: 2px 6px;
+  cursor: pointer;
+  background: var(--color-danger);
+  border: none;
+  color: #fff;
+}
+
+.sp-delete-no {
+  font-size: 10px;
+  padding: 2px 6px;
+  cursor: pointer;
+  background: rgba(255, 255, 255, 0.2);
+  border: none;
+  color: #fff;
+}
+
+.sp-thumb-remove {
+  position: absolute;
+  top: 3px;
+  right: 3px;
+  width: 18px;
+  height: 18px;
+  background: rgba(0, 0, 0, 0.55);
+  border: none;
+  color: #fff;
+  cursor: pointer;
+  font-size: 13px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0;
+  line-height: 1;
+}
